backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
python-docx==1.1.2
pdfplumber==0.11.7
openpyxl==3.1.5
pytesseract==0.3.13
Pillow==11.3.0
rapidfuzz==3.13.0
sentence-transformers==5.1.0
spacy==3.8.7
numpy==2.3.2
pydantic==2.11.7
orjson==3.11.1
